Refactor across multiple files
