+++ Sources/HashableMacroMacros/Diagnostics/HashableMacroFixItMessage.swift
@@ -1,0 +1,12 @@
+import SwiftDiagnostics
+
+struct HashableMacroFixItMessage: FixItMessage {
+    let fixItID: MessageID
+    let message: String
+
+    init(id: String, message: String) {
+        fixItID = MessageID.makeHashableMacroMessageID(id: id)
+        self.message = message
+    }
+}
+
