+++ Sources/HashableMacroMacros/Diagnostics/MessageID+HashableMacro.swift
@@ -1,0 +1,7 @@
+import SwiftDiagnostics
+
+extension MessageID {
+    static func makeHashableMacroMessageID(id: String) -> MessageID {
+        MessageID(domain: "uk.josephduffy.HashableMacro", id: id)
+    }
+}
