+++ Sources/HashableMacroMacros/Macros/HashableMacro+Hashable.swift
@@ -1,0 +1,222 @@
+import Foundation
+import SwiftDiagnostics
+import SwiftSyntax
+import SwiftSyntaxMacros
+
+extension HashableMacro {
+    static func expansionForHashable(
+        of node: AttributeSyntax,
+        providingMembersOf declaration: some DeclGroupSyntax,
+        in context: some MacroExpansionContext,
+        propertiesToHash: [TokenSyntax]
+    ) -> DeclSyntax {
+        var finalHashInto = true
+
+        if let arguments = node.arguments?.as(LabeledExprListSyntax.self) {
+            for argument in arguments {
+                guard let label = argument.label else { continue }
+                switch label.trimmed.text {
+                case "finalHashInto":
+                    guard let expression = argument.expression.as(BooleanLiteralExprSyntax.self) else { continue }
+                    switch expression.literal.tokenKind {
+                    case .keyword(.true):
+                        finalHashInto = true
+                    case .keyword(.false):
+                        finalHashInto = false
+                    default:
+                        break
+                    }
+                default:
+                    break
+                }
+            }
+        }
+
+        let baseModifiers = declaration.modifiers.filter({ modifier in
+            switch (modifier.name.tokenKind) {
+            case .keyword(.public):
+                return true
+            case .keyword(.internal):
+                return true
+            case .keyword(.fileprivate):
+                return true
+            case .keyword(.private):
+                // The added functions should never be private
+                return false
+            default:
+                return false
+            }
+        })
+
+        var hashFunctionModifiers = baseModifiers
+        if finalHashInto, declaration.is(ClassDeclSyntax.self) {
+            hashFunctionModifiers.append(
+                DeclModifierSyntax(name: .keyword(.final))
+            )
+        }
+
+        let hashFunctionSignature = FunctionSignatureSyntax(
+            parameterClause: FunctionParameterClauseSyntax(
+                parameters: [
+                    FunctionParameterSyntax(
+                        firstName: .identifier("into"),
+                        secondName: .identifier("hasher"),
+                        type: AttributedTypeSyntax(
+                            specifier: .keyword(.inout),
+                            baseType: TypeSyntax(stringLiteral: "Hasher")
+                        )
+                    ),
+                ]
+            )
+        )
+
+        let hashFunctionBody = CodeBlockSyntax(
+            statements: CodeBlockItemListSyntax(itemsBuilder: {
+                for propertyToken in propertiesToHash {
+                    FunctionCallExprSyntax(
+                        callee: MemberAccessExprSyntax(
+                            base: DeclReferenceExprSyntax(baseName: "hasher"),
+                            name: .identifier("combine")
+                        ),
+                        argumentList: {
+                            LabeledExprSyntax(
+                                expression: MemberAccessExprSyntax(
+                                    base: DeclReferenceExprSyntax(baseName: .keyword(.`self`)),
+                                    name: propertyToken
+                                )
+                            )
+                        }
+                    )
+                }
+            })
+        )
+
+        let hashFunction = FunctionDeclSyntax(
+            modifiers: hashFunctionModifiers,
+            name: .identifier("hash"),
+            signature: hashFunctionSignature,
+            body: hashFunctionBody
+        )
+
+        return DeclSyntax(hashFunction)
+    }
+
+    static func expansionForEquals(
+        of node: AttributeSyntax,
+        providingMembersOf declaration: some DeclGroupSyntax,
+        in context: some MacroExpansionContext,
+        propertiesToHash: [TokenSyntax]
+    ) throws -> DeclSyntax {
+        guard let namedDeclaration = declaration as? NamedDeclSyntax else {
+            throw HashableMacroDiagnosticMessage(
+                id: "not-named-declaration",
+                message: "'@Hashable' can only be applied to named declarations",
+                severity: .error
+            )
+        }
+
+        let baseModifiers = declaration.modifiers.filter({ modifier in
+            switch (modifier.name.tokenKind) {
+            case .keyword(.public):
+                return true
+            case .keyword(.internal):
+                return true
+            case .keyword(.fileprivate):
+                return true
+            case .keyword(.private):
+                // The added functions should never be private
+                return false
+            default:
+                return false
+            }
+        })
+
+        let equalsFunctionSignature = FunctionSignatureSyntax(
+            parameterClause: FunctionParameterClauseSyntax(
+                parameters: [
+                    FunctionParameterSyntax(
+                        firstName: .identifier("lhs"),
+                        type: TypeSyntax(stringLiteral: namedDeclaration.name.text),
+                        trailingComma: .commaToken()
+                    ),
+                    FunctionParameterSyntax(
+                        firstName: .identifier("rhs"),
+                        type: TypeSyntax(stringLiteral: namedDeclaration.name.text)
+                    ),
+                ]
+            ),
+            returnClause: ReturnClauseSyntax(
+                type: IdentifierTypeSyntax(name: .identifier("Bool"))
+            )
+        )
+
+        var comparisons: InfixOperatorExprSyntax?
+
+        for propertyToken in propertiesToHash {
+            let comparison = InfixOperatorExprSyntax(
+                leftOperand: MemberAccessExprSyntax(
+                    base: DeclReferenceExprSyntax(
+                        baseName: .identifier("lhs")
+                    ),
+                    declName: DeclReferenceExprSyntax(
+                        baseName: propertyToken
+                    )
+                ),
+                operator: BinaryOperatorExprSyntax(
+                    operator: .binaryOperator("==")
+                ),
+                rightOperand: MemberAccessExprSyntax(
+                    base: DeclReferenceExprSyntax(
+                        baseName: .identifier("rhs")
+                    ),
+                    declName: DeclReferenceExprSyntax(
+                        baseName: propertyToken
+                    )
+                )
+            )
+
+            if let existingComparisons = comparisons {
+                comparisons = InfixOperatorExprSyntax(
+                    leftOperand: existingComparisons,
+                    operator: BinaryOperatorExprSyntax(
+                        leadingTrivia: .newline.appending(Trivia.spaces(8)),
+                        operator: .binaryOperator("&&")
+                    ),
+                    rightOperand: comparison
+                )
+            } else {
+                comparisons = comparison
+            }
+        }
+
+        let equalsBody = CodeBlockSyntax(
+            statements: CodeBlockItemListSyntax(itemsBuilder: {
+                if let comparisons {
+                    ReturnStmtSyntax(
+                        leadingTrivia: .spaces(4),
+                        expression: comparisons
+                    )
+                } else {
+                    ReturnStmtSyntax(
+                        leadingTrivia: .spaces(4),
+                        expression: BooleanLiteralExprSyntax(booleanLiteral: true)
+                    )
+                }
+            })
+        )
+
+        var equalsFunctionModifiers = baseModifiers
+        equalsFunctionModifiers.append(
+            DeclModifierSyntax(name: .keyword(.static))
+        )
+
+        let equalsFunction = FunctionDeclSyntax(
+            modifiers: equalsFunctionModifiers,
+            name: .identifier("=="),
+            signature: equalsFunctionSignature,
+            body: equalsBody
+        )
+
+        return DeclSyntax(equalsFunction)
+    }
+}
